Oops, should be floats

+++ idg/idgmsgridder.cpp
@@ -84,9 +84,10 @@ void IdgMsGridder::Invert() {
       max_w = std::max(max_w, msDataVector[i].maxWWithFlags);
     }
 
-    const double shiftl = PhaseCentreDL();
-    const double shiftm = PhaseCentreDM();
-    const double shiftp = sqrt(1.0 - shiftl * shiftl - shiftm * shiftm) - 1.0;
+    const float shiftl = PhaseCentreDL();
+    const float shiftm = PhaseCentreDM();
+    const float shiftp =
+        std::sqrt(1.0 - shiftl * shiftl - shiftm * shiftm) - 1.0;
     _bufferset->init(width, _actualPixelSizeX, max_w + 1.0, shiftl, shiftm,
                      shiftp, _options);
     Logger::Debug << "IDG subgrid size: " << _bufferset->get_subgridsize()
@@ -296,9 +297,10 @@ void IdgMsGridder::Predict(ImageF image) {
       max_w = std::max(max_w, msDataVector[i].maxWWithFlags);
     }
 
-    const double shiftl = PhaseCentreDL();
-    const double shiftm = PhaseCentreDM();
-    const double shiftp = sqrt(1.0 - shiftl * shiftl - shiftm * shiftm) - 1.0;
+    const float shiftl = PhaseCentreDL();
+    const float shiftm = PhaseCentreDM();
+    const float shiftp =
+        std::sqrt(1.0 - shiftl * shiftl - shiftm * shiftm) - 1.0;
     _bufferset->init(width, _actualPixelSizeX, max_w + 1.0, shiftl, shiftm,
                      shiftp, _options);
     _bufferset->set_image(_image.data(), do_scale);
